Twitter hotfix Twitter hotfix now added to the bot, don't need to start a new process
